split `lib/lc/client.ts` into separate files
